- renamed package job_manager to job_management
